Update game to handle using the Blender model
Since the [model has to be added to the scene in the callback run after loading the model from JSON](http://stackoverflow.com/questions/13655092/using-three-js-jsonloader), I had to move some stuff around to pass the Scene object to the ship's constuctor through the player.  I also moved creating the ship to a method on the player, which makes more sense since in the future we'll probably support multiple ship types and stuff.  It also removes creating a ship from the "application" itself, which makes sense one we're trying to support multiple players.